pyproject.toml
[tool.poetry]
name = "azfunctionsmonitor"
version = "0.1.112"
description = "Loging helper for Azure Monitor and Open Census"
authors = ["Shawn Cicoria <github@cicoria.com>"]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://github.com/cicorias/az-monitor-py-logging-pkg"
documentation = "https://github.com/cicorias/az-monitor-py-logging-pkg"
keywords = ["azure", "monitor", "cloud"]
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Topic :: Software Development :: Libraries :: Python Modules"
]

[tool.poetry.dependencies]
python = "^3.7"
azure-functions = "^1.6.0"
opencensus-ext-azure = "^1.0.7"

[tool.poetry.dev-dependencies]
flake8 = "^3.8"
pytest = "^6.2"
black = {version = "^20.8b1", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
